+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Breaking Changes
+
+- **message:** Rename `ParseRawMessage` to `ParseRaw` for consistency
+
 ## [0.11.0] - 2025-01-09
 
 ### Breaking Changes